Moving Footer to the features folder
