+++ amundsen_application/static/js/features/Footer/styles.scss
